+++ demo/forms.py
@@ -152,3 +152,20 @@ Use your favorite Browser dev tool to inspect it.
     person = fields.AgnocompleteField(
         'AutocompleteUrlErrors',
         help_text='will never find anybody')
+
+
+class UrlProxyWithExtraForm(forms.Form):
+    help_text = UrlProxyFormMixin.help_text + """
+
+Please note that we're adding an extra argument to the search: `special`.
+"""
+
+    person_simple = fields.AgnocompleteField(
+        'AutocompleteUrlSimple',
+        help_text='Simple search using GET',
+    )
+
+    person_simple_extra = fields.AgnocompleteField(
+        'AutocompleteUrlSimpleWithExtra',
+        help_text='Simple search using GET, exploiting the "special" argument',
+    )