+++ demo/tests/__init__.py
@@ -34,7 +34,7 @@ class LoaddataLiveTestCase(LoaddataMixin, LiveServerTestCase):
 class RegistryTestGeneric(LoaddataTestCase):
 
     def _test_registry_keys(self, keys):
-        self.assertEqual(len(keys), 21)
+        self.assertEqual(len(keys), 22)
         self.assertIn("AutocompleteColor", keys)
         self.assertIn("AutocompleteColorExtra", keys)
         self.assertIn("AutocompletePerson", keys)
@@ -62,6 +62,7 @@ class RegistryTestGeneric(LoaddataTestCase):
         self.assertIn("AutocompleteUrlSimpleAuth", keys)
         self.assertIn("AutocompleteUrlHeadersAuth", keys)
         self.assertIn("AutocompleteUrlErrors", keys)
+        self.assertIn("AutocompleteUrlSimpleWithExtra", keys)
 
 
 class MockRequestUser(object):