+++ demo/views.py
@@ -19,6 +19,7 @@ from .forms import (
     PersonContextTagModelForm,
     UrlProxyForm, UrlProxyConvertForm,
     UrlProxyAuthForm, UrlProxyErrors,
+    UrlProxyWithExtraForm,
 )
 from .autocomplete import HiddenAutocomplete
 from .models import PersonTag
@@ -197,6 +198,12 @@ class UrlProxyErrorsView(AutoView):
     template_name = "selectize.html"
 
 
+class UrlProxyWithExtraView(AutoView):
+    form_class = UrlProxyWithExtraForm
+    title = 'Authenticated URLs, returned normal data'
+    template_name = "selectize.html"
+
+
 index = IndexView.as_view()
 filled_form = FilledFormView.as_view()
 search_context = SearchContextFormView.as_view()
@@ -221,3 +228,4 @@ url_proxy_simple = UrlProxySimpleView.as_view()
 url_proxy_convert = UrlProxyConvertView.as_view()
 url_proxy_auth = UrlProxyAuthView.as_view()
 url_proxy_errors = UrlProxyErrorsView.as_view()
+url_proxy_with_extra = UrlProxyWithExtraView.as_view()